Add View file and move code.
